refactor: move inspection to component, minor UI
