+++ packages/hoppscotch-common/src/components/http/ResponseMeta.vue
@@ -2,34 +2,15 @@
   <div
     class="sticky top-0 z-10 flex items-center justify-center flex-shrink-0 p-4 overflow-auto overflow-x-auto bg-primary whitespace-nowrap"
   >
-    <div
-      v-for="(inspector, index) in tabResults"
-      :key="index"
+    <AppInspection
+      :inspection-results="tabResults"
       :class="[
         response === null || response?.type === 'network_fail'
           ? 'absolute left-2 top-2'
           : 'mr-2',
       ]"
-    >
-      <div
-        v-if="inspector.isApplicable"
-        class="flex justify-center items-center"
-      >
-        <tippy ref="options" interactive theme="popover">
-          <div class="flex justify-center items-center flex-1 felx-col">
-            <HoppButtonSecondary
-              :icon="inspector.icon"
-              :class="severityColor(inspector.severity)"
-            />
-          </div>
-          <template #content>
-            <span v-if="inspector.text.type === 'text'">
-              {{ inspector.text.text }}
-            </span>
-          </template>
-        </tippy>
-      </div>
-    </div>
+    />
+
     <AppShortcutsPrompt v-if="response == null" class="flex-1" />
     <div v-else class="flex flex-col flex-1">
       <div
@@ -160,18 +141,6 @@ const statusCategory = computed(() => {
   return findStatusGroup(props.response.statusCode)
 })
 
-const severityColor = (severity: number) => {
-  switch (severity) {
-    case 1:
-      return "text-green-500"
-    case 2:
-      return "text-yellow-500"
-    case 3:
-      return "text-red-500"
-    default:
-      return "text-gray-500"
-  }
-}
 const inspectionService = useService(InspectionService)
 
 const allTabResults = inspectionService.tabs
